Hotfix for the race condition in reading/writing the config json file.
See https://github.com/fchollet/keras/issues/3498

+++ keras/backend/__init__.py
@@ -37,13 +37,14 @@ if os.path.exists(_config_path):
     set_epsilon(_epsilon)
     _BACKEND = _backend
 
-# save config file
-_config = {'floatx': floatx(),
-           'epsilon': epsilon(),
-           'backend': _BACKEND,
-           'image_dim_ordering': image_dim_ordering()}
-with open(_config_path, 'w') as f:
-    f.write(json.dumps(_config, indent=4))
+if not os.path.exists(_config_path):
+    # save config file
+    _config = {'floatx': floatx(),
+               'epsilon': epsilon(),
+               'backend': _BACKEND,
+               'image_dim_ordering': image_dim_ordering()}
+    with open(_config_path, 'w') as f:
+        f.write(json.dumps(_config, indent=4))
 
 if 'KERAS_BACKEND' in os.environ:
     _backend = os.environ['KERAS_BACKEND']
